requirements.txt
python-dotenv==1.0.1
langgraph==0.2.54
langchain-google-genai==2.0.6
tavily-python==0.5.0
pydantic-ai==0.0.11